Throw more specific errors
Scalaqlite throws rather inscrutable errors that return the result code,
but no description of the corresponding error. I remedy this by copying
the documentation for all of the result codes and adding it to the
exception. I also switch to throwing and SqlException instead of using
sys.error (which throws a RuntimeException). To improve pattern matching
support, I make SqlException a case class.

Bonus: switch to string interpolation in a few places